DECISIONS-CONTROLL-STATEMENT
A decision control statement (also known as a conditional statement) is a programming construct that allows the program to make decisions based on conditions. In Python, the most common decision control statements are if, else-if, and else.

if is used to check if a condition is true.
else-if (else if) is used if the first condition is false and you want to check another condition.
else is used if none of the conditions are true.